Improve handling of null reviewDecision
After more research into when a `null` reviewDecision can occur, I determined
we can be a bit smarter about how we assign the review status. Whenever
the reviewDecision is `null` it means the repository has not enabled
the `Require pull request reviews before merging` branch protection feature.
In this scenario any 'approving' review is sufficient to determine if the
cope was approved. Thus, whenever the `reviewDecision` field is `null`,
we check if any reviews have the `APPROVED` state, and set the `reviewDecision`
to `APPROVED` if so, otherwise we set it to the default status of `UNKNOWN`.