Switch from usings reviewDecision to reviews (#231)
After some investigation, it was determined the `reviewDecision` field
is unreliable as it can be set to `null` after a PR is merged. For
example, if a user un-resolves a conversation on PR post-merge, then the
`reviewDecision` will go from `APPROVED` to `null`.

This change removes the `reviewDecision` field and instead uses the
underlying review data of a PR to determine the approval status of a
commit.