Remove DeliveryEventExists check on webhook flow (#407)
We were running into this error when processing webhooks:

```
failed to execute DeliveryEventExists: failed to make read request to BigQuery: googleapi: Error 403: Quota exceeded: Your project_and_region exceeded quota for max number of jobs that can be queued per project. For more information, see https://cloud.google.com/bigquery/docs/troubleshoot-quotas, quotaExceeded
```

This is because we were submitting too many jobs checking if every event
already exists. We already do this on the query side (we filter for
unique events) so this check is redundant. There is no way to increase
this quota (though switching to batch priority could improve throughput
at the cost of latency which may be unacceptable).

I am leaving this check in for the retry flow as it is expected to be at
a much lower QPS.

Resolves #406