feat: refactored retry service into a cloud run job (#417)
closes #413

This PR refactors the retry service from a Cloud Run service to a Cloud
Run job. This is a more appropriate and efficient model for the retry
mechanism, which runs periodically to find and retry failed webhook
deliveries.

Key changes include:

- Application Logic: The HTTP server code in pkg/retry has been removed
and replaced with a ExecuteJob function that encapsulates the retry
logic. The command-line interface has been updated to execute this job.
- Terraform: The infrastructure in terraform/service_retry.tf has been
updated to deploy a google_cloud_run_v2_job instead of a
google_cloud_run_service. This includes the necessary IAM roles and a
google_cloud_scheduler_job to trigger the job.
- GitHub Actions: The CI/CD workflow in .github/workflows/ci.yml has
been updated to remove the deployment of the old retry service. This
workflow will be reworked to deploy the CR job once the terraform can be
applied in those environments.
- Testing: Unit tests have been updated to reflect the new job-based
architecture.

This change simplifies the architecture and reduces the resource
footprint of the retry mechanism.

want_lgtm=all