fixing ENV variable names (#426)
Fixing Secret Environment variable so container can access secret and
forcing latest version of secret so if it's updated it won't break. The
default was not enough since the container have already a version in the
image so it won't replace the version if there is no value beside
default.